Clean Kirin db at each test, only for GuichetUnique
Also reactivate test_kirin_cots_trip_add_new_stop_point_in_the_middle that was not chaining well with previous test.
